requirements.txt
# fastapi[all]==0.115.0
# pydantic==2.9.2
pydantic[email]
# uvicorn==0.31.0
jinja2==3.1.4
# pydantic_settings==2.5.2
# aiosqlite==0.20.0
# alembic==1.13.3
# SQLAlchemy==2.0.35
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
# loguru==0.7.2
aiogram==3.17.0
aiogram_dialog==2.3.1
sqlalchemy==2.0.38
pydantic==2.10.6
pydantic_settings==2.7.1
aio-pika==9.5.4
faststream==0.5.34
loguru==0.7.3
aiosqlite==0.21.0
alembic==1.14.1
pytz==2025.1
apscheduler==3.11.0
fastapi==0.115.8
uvicorn==0.34.0